requirements.txt
colorlog==6.7.0
arrow==1.2.3
atlassian-python-api==3.39.0
beartype==0.14.1
beautifulsoup4==4.12.2
cryptography==38.0.4  # don't change it - dependence to Django
colorama==0.4.6
dominate==2.8.0
elasticsearch==7.17.0 # don't change it - dependence to elasticsearch-dsl
elasticsearch-dsl==7.4.1
github3.py==4.0.1
ipaddr==2.2.0
jira==3.5.1
matplotlib==3.7.1
netaddr==0.8.0
haversine==2.7.0 # don't change it - 2.8 failing to install
msgpack==1.0.5
nested-lookup==0.2.25
numpy==1.25.0
nrarfcn==2.4.0
openpyxl==3.1.2
pandas==2.0.2
paramiko==3.2.0
Pillow==9.5.0
plotly==5.15.0
pysnmp==4.4.12
pytrafficgen==4.0.1
pyasn1==0.4.8  # don't change it - 0.5.0 not working well with pysnmp
PyYAML==6.0
requests==2.31.0
robotframework==6.1
robotframework-excellib==2.0.1
robotframework-metrics==3.3.3
robotframework-sshlibrary==3.8.0
robotframework-scplibrary==1.2.0
robotframework-websocketclient==1.3.0
robotframework-jsonlibrary==0.5
robotframework-requests==0.9.5
robotframework-tidy==4.3.0
robotframework-seleniumlibrary==6.1.0
robotframework-screencaplibrary==1.6.0
robotframework-appiumlibrary==2.0.0
scp==0.14.5
termcolor==2.3.0
touca==1.8.7
vmwc==1.0.6
xlrd==2.0.1
xlwings==0.30.10
jsonref==1.1.0
pandasticsearch==0.6.0
DateTimeRange==2.1.0
ipython==8.14.0
altair==5.0.1
jsonschema==4.17.3
pymodbus==2.5.3  # Do not upgrade version, affecting automation keyword
python-jenkins==1.8.0
python-dotenv==1.0.0
pygit2==1.12.1
GitPython==3.1.31
jenkinsapi==0.3.13
PyPDF2==3.0.1
tenable-io==1.13.1
xmltodict==0.13.0
PyGithub==1.59.0
scapy==2.5.0
nest_asyncio==1.5.6
# mac-vendor-lookup==0.1.12 # this verion not supported on python 3.11
manuf==1.1.5
simple-websocket-server==0.4.0  # Do not upgrade version
websocket-client==0.53.0  # Do not upgrade version
websockets==11.0.3
imutils==0.5.4
scikit-image==0.21.0
opencv-python==4.7.0.72
pyshark==0.6
zeep==4.2.1